chore: Overwrite JitPack dependencies for pull request builds
This change in GitHub workflow definition enables PRs to use new/changed utility classes in module camel-kamelets-utils. The workflow overwrites the JitPack dependency coordinates in all Kamelets to point to the actual PR head ref instead of base (e.g. main-SNAPSHOT).

This ensures that the automated tests use the utility classes from that specific PR instead of those from the base branch (e.g. main-SNAPSHOT). So pull request builds are able to have a green GitHub CI workflow even if the base branch is not aware of the utility class before the merge.